+++ src/datasource-zabbix/module.ts
@@ -2,15 +2,18 @@ import { loadPluginCss } from '@grafana/runtime';
 import { ZabbixDatasource } from './datasource';
 import { ZabbixQueryController } from './query.controller';
 import { ZabbixDSConfigController } from './config.controller';
+import { ZabbixVariableQueryEditor } from './components/VariableQueryEditor';
 import './zabbixAlerting.service.js';
 import './add-metric-function.directive';
 import './metric-function-editor.directive';
 
-class ZabbixQueryOptionsController {}
-ZabbixQueryOptionsController.templateUrl = 'datasource-zabbix/partials/query.options.html';
+class ZabbixQueryOptionsController {
+  static templateUrl = 'datasource-zabbix/partials/query.options.html';
+}
 
-class ZabbixAnnotationsQueryController {}
-ZabbixAnnotationsQueryController.templateUrl = 'datasource-zabbix/partials/annotations.editor.html';
+class ZabbixAnnotationsQueryController {
+  static templateUrl = 'datasource-zabbix/partials/annotations.editor.html';
+}
 
 ZabbixQueryController.templateUrl = 'datasource-zabbix/partials/query.editor.html';
 ZabbixDSConfigController.templateUrl = 'datasource-zabbix/partials/config.html';
@@ -25,5 +28,6 @@ export {
   ZabbixDSConfigController as ConfigCtrl,
   ZabbixQueryController as QueryCtrl,
   ZabbixQueryOptionsController as QueryOptionsCtrl,
-  ZabbixAnnotationsQueryController as AnnotationsQueryCtrl
+  ZabbixAnnotationsQueryController as AnnotationsQueryCtrl,
+  ZabbixVariableQueryEditor as VariableQueryEditor,
 };