+++ src/panel-triggers/types.ts
@@ -1,7 +1,7 @@
 export interface ProblemsPanelOptions {
   schemaVersion: number;
   datasources: any[];
-  targets: Map<string, ProblemsPanelTarget>;
+  targets: ProblemsPanelTarget[];
   // Fields
   hostField?: boolean;
   hostTechNameField?: boolean;
@@ -62,6 +62,7 @@ export interface ProblemsPanelTarget {
   proxy: {
     filter: string
   };
+  datasource: string;
 }
 
 export interface TriggerSeverity {
@@ -192,7 +193,7 @@ export interface RTRow<T> {
   /** true if this row was produced by a pivot */
   groupedByPivot?: boolean;
   /** any sub rows defined by the `subRowKey` prop */
-  subRows?: boolean;
+  subRows?: any[];
 }
 
 export interface RTCell<T> extends RTRow<T> {