+++ src/test-setup/jest-setup.js
@@ -82,10 +82,6 @@ jest.mock('grafana/app/core/config', () => {
 
 jest.mock('jquery', () => 'module not found', {virtual: true});
 
-jest.mock('@grafana/ui');
-
-jest.mock('@grafana/runtime');
-
 // Required for loading angularjs
 let dom = new JSDOM('<html><head><script></script></head><body></body></html>');
 // Setup jsdom